pyproject.toml
[project]
name = "itexus-spec-kit"
version = "0.1.3"
description = "Itexus accelerator for spec-driven AI delivery on top of github/spec-kit"
requires-python = ">=3.10"
dependencies = [
  "PyYAML==6.0.2",
]

[project.optional-dependencies]
dev = [
  "ruff==0.12.7",
  "mypy==1.17.1",
  "pip-audit==2.9.0",
]

[tool.ruff]
line-length = 120
target-version = "py310"
exclude = [
  "dist",
  ".venv",
  ".pytest_cache",
]

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
ignore_missing_imports = true
exclude = "(^dist/|^\\.venv/|^\\.pytest_cache/)"
